requirements.txt
# Core
pymupdf==1.24.9
python-dotenv==1.0.1
streamlit==1.38.0
groq==0.11.0

# Vector DB + Embeddings
chromadb==0.5.5
sentence-transformers==3.0.1

# Optional (table extraction / OCR) - install only if you need them:
# camelot-py==0.11.0
# tabula-py==2.9.3
# pytesseract==0.3.10
# pillow==10.4.0
